+++ TrickyTowerDefense/Assets/Scripts/Unit/EnemyHealth.cs
@@ -6,7 +6,7 @@ public class EnemyHealth : MonoBehaviour, IHitable
 {
     EnemyScript enemyScript = null;
 
-    public void Hit(int damage)
+    public void Hit(float damage)
     {
         enemyScript.Data.Damage(damage);
     }
